requirements.txt
asgiref==3.3.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.3
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.7
Django==3.0
django-allauth==0.44.0
django-cors-headers==3.5.0
django-decouple==2.1
django-filter==2.4.0
djangorestframework==3.12.1
idna==2.10
oauthlib==3.1.1
paramiko==2.7.2
Pillow==8.0.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
PyNaCl==1.4.0
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.5
